legacy gcp cloud function/upload_To_bucket/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-storage-control==1.*